* further magic for the Bloom filter

+++ lola2/src/Stores/BloomStore.h
@@ -7,10 +7,10 @@ class BloomStore : public Store
 {
     private:
         std::bitset<BLOOM_FILTER_SIZE> *filter;
-        inline uint64_t hash_sdbm() const;
-        inline uint64_t hash_fnv() const;
+        inline unsigned int hash_sdbm() const;
+        inline unsigned int hash_fnv() const;
 
-        uint64_t *hash_values;
+        unsigned int *hash_values;
         size_t hash_functions;
 
     public:
